Fix transcription
