+++ packages/plugins/experimental/plugin-transcription/src/hooks/index.ts
@@ -2,8 +2,6 @@
 // Copyright 2025 DXOS.org
 //
 
-export * from './transcription';
-
 export * from './useAudioFile';
 export * from './useAudioTrack';
 export * from './useIsSpeaking';
